WIP experiment replacing defrecord with deftype
This avoids issues when printing a value containing a mocked connection.
This also avoids cryptic errors when calling datomic.api/db on a mocked
connection which is only a problem for beginners, but still, this is a
difference compared to a normal datomic connection.

Problem: This is potentially  a breaking change. Maybe we can get those
benefits while keeping the record?